Enable PodSecurityPolicy by default
This commit enables the PodSecurityPolicy (PSP) admission controller in
Tectonic by amending it to the admission-controller flag of the
kube-apiserver. The --use-service-account-credentials flag has been added to the
kube-controller-manager, ensuring it runs with all individual
ServiceAccounts provided by k8s [0].

A 'permissive' and 'limited' PSP have been added for use by the
kube-controller-manager. The permissive PSP has no restrictions on
pods to create. The limited policy is the same as permissive with
the following added restrictions.

  - Pods may not run as privileged [1]
  - Pods may not run using host networking [2]

A ClusterRole allowing the verb 'use' for each PSP resource ('permissive',
'limited') has been created for binding to. Binding to means the
ServiceAccount will be able to resolve against that PSP when attempting
to create a pod. A ClusterRoleBinding has been created for the
limited PSP ClusterRole. Making all pod creations in all namespaces
require to match the rules set in the limited PSP. RoleBindings have
been created in kube-system and tectonic-system for the permissive PSP.
Making all pods created in the kube-system or tectonic-system namespace
able to resolve against the permissive PSP, containing no restrictions.
All mentioned bindings include the Subjects for every ServiceAccount
that uses the 'create' verb on the 'pod' resource. The following
ServiceAccounts (from the kube-system namespace) are bound.

  - replicaset-controller
  - replication-controller
  - job-controller
  - daemon-set-controller
  - statefulset-controller

This change will prevent users from being able to create pods in
namespaces as their user account won't have access to 'use' any PSP.
They will have to rely on high-level constructs, which they normally
would, to create pods such as StatefulSets or Deployments. When rights
are desired beyond that of the limited PSP, admins will need to add new
PSPs and likely change RBAC rules [3].

Resolves: https://github.com/coreos/tectonic-installer/issues/2101

[0]: https://kubernetes.io/docs/admin/authorization/rbac/#controller-roles
[1]: https://kubernetes.io/docs/concepts/workloads/pods/pod/#privileged-mode-for-pod-containers
[2]: https://kubernetes.io/docs/concepts/policy/pod-security-policy/#host-network
[3]: https://github.com/kubernetes/examples/blob/master/staging/podsecuritypolicy/rbac/README.md